requirements.txt
Django==3.2.12
djangorestframework==3.12.4
django-jenkins==0.110.0
# pep8==1.7.1
# pyflakes==2.4.0